+++ app/Http/Resources/LinkResource.php
@@ -13,7 +13,11 @@ class LinkResource extends JsonResource
             'title' => $this->title,
             'content' => $this->content,
             'url' => $this->url,
+            'http_status' => $this->http_status,
+            'http_status_color' => $this->getHttpStatusColor($this->http_status),
+            'http_checked_at_formated' => $this->http_checked_at ? $this->http_checked_at->diffForHumans() : '',
             'permalink' => $this->permalink,
+            'is_watched' => $this->is_watched,
             'is_private' => $this->post->is_private,
             'is_pinned' => $this->post->is_pinned,
             'preview' => $this->preview,
@@ -32,5 +36,21 @@ class LinkResource extends JsonResource
                'url_download' => route('link.archive.download', $this->id),
             ])
         ];
+    }
+
+    protected function getHttpStatusColor($status): ?string
+    {
+        $range = (int)floor((int)$status / 100);
+
+        switch ($range) {
+            case 3:
+                return 'info';
+            case 4:
+                return 'warning';
+            case 5:
+                return 'danger';
+            default:
+                return null;
+        }
     }
 }