+++ resources/views/manage/links.blade.php
@@ -1,0 +1,6 @@
+@extends('layouts.manage')
+
+@section('content')
+    <health-checks :enabled="{{ $enabled ? 'true' : 'false' }}"
+                   :stats="{{ json_encode($stats) }}"></health-checks>
+@endsection